+++ test-programs/build-all.sh
@@ -1,0 +1,4 @@
+#!/usr/bin/env bash
+set -eux -o pipefail
+
+( cd nop; make )