+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `Error::TraceeDied` variant, `Error::tracee_died()` method to improve error ergonomics
+- x86 `Tracee` methods for reading and writing debug registers
 
 ### Changed